requirements.txt
flask==2.1.2
pillow
werkzeug==2.0.3
